// clients/typescript-client/src/generated/jetstream_protos/protos/jetstream.ts
export interface SubscribeRequest {
  transactions: { [key: string]: SubscribeRequestFilterTransactions };
  accounts: { [key: string]: SubscribeRequestFilterAccounts };
  ping?: SubscribeRequestPing | undefined;
}

export interface SubscribeRequest_TransactionsEntry {
  key: string;
  value?: SubscribeRequestFilterTransactions | undefined;
}

export interface SubscribeRequest_AccountsEntry {
  key: string;
  value?: SubscribeRequestFilterAccounts | undefined;
}

export interface SubscribeRequestFilterTransactions {
  accountInclude: string[];
  accountExclude: string[];
  accountRequired: string[];
}

export interface SubscribeRequestFilterAccounts {
  account: string[];
  owner: string[];
  filters: SubscribeRequestFilterAccountsFilter[];
}
// ...
export interface SubscribeRequestPing {
  id: number;
}

/** ============= Subscribe Updates ============= */
export interface SubscribeUpdate {
  filters: string[];
  createdAt?: Date | undefined;
  transaction?: SubscribeUpdateTransaction | undefined;
  account?: SubscribeUpdateAccount | undefined;
  ping?: SubscribeUpdatePing | undefined;
  pong?: SubscribeUpdatePong | undefined;
}
// ...
export interface SubscribeUpdatePing {
}

export interface SubscribeUpdatePong {
  id: number;
}

export interface MessageAddressTableLookup {
  accountKey: Buffer;
  writableIndexes: Buffer;
  readonlyIndexes: Buffer;
}

export interface CompiledInstruction {
  programIdIndex: number;
  accounts: Buffer;
  data: Buffer;
}

/** ============= Parsed Instruction Messages ============= */
export interface SubscribeParsedRequest {
  ping?: SubscribeRequestPing | undefined;
}
// ...
export interface SubscribeUpdateParsedTransaction {
  signature: Buffer;
  slot: number;
  account?: SubscribeUpdateAccount | undefined;
  recentBlockhash: Buffer;
  signatures: Buffer[];
  instructions: Instruction[];
}

/** ============= Non-streaming Methods ============= */
export interface PingRequest {
  count: number;
}

export interface PongResponse {
  count: number;
}

export interface GetVersionRequest {
}

export interface GetVersionResponse {
  version: string;
}

export interface GetSlotResponse {
  slot: number;
}

function createBaseSubscribeRequest(): SubscribeRequest {
  return { transactions: {}, accounts: {}, ping: undefined };
}

export const SubscribeRequest = {
  encode(message: SubscribeRequest, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    Object.entries(message.transactions).forEach(([key, value]) => {
      SubscribeRequest_TransactionsEntry.encode({ key: key as any, value }, writer.uint32(10).fork()).ldelim();
    });
    Object.entries(message.accounts).forEach(([key, value]) => {
      SubscribeRequest_AccountsEntry.encode({ key: key as any, value }, writer.uint32(18).fork()).ldelim();
    });
    if (message.ping !== undefined) {
      SubscribeRequestPing.encode(message.ping, writer.uint32(34).fork()).ldelim();
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): SubscribeRequest {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseSubscribeRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 10) {
            break;
          }

          const entry1 = SubscribeRequest_TransactionsEntry.decode(reader, reader.uint32());
          if (entry1.value !== undefined) {
            message.transactions[entry1.key] = entry1.value;
          }
          continue;
        case 2:
          if (tag !== 18) {
            break;
          }

          const entry2 = SubscribeRequest_AccountsEntry.decode(reader, reader.uint32());
          if (entry2.value !== undefined) {
            message.accounts[entry2.key] = entry2.value;
          }
          continue;
        case 4:
          if (tag !== 34) {
            break;
          }

          message.ping = SubscribeRequestPing.decode(reader, reader.uint32());
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): SubscribeRequest {
    return {
      transactions: isObject(object.transactions)
        ? Object.entries(object.transactions).reduce<{ [key: string]: SubscribeRequestFilterTransactions }>(
          (acc, [key, value]) => {
            acc[key] = SubscribeRequestFilterTransactions.fromJSON(value);
            return acc;
          },
          {},
        )
        : {},
      accounts: isObject(object.accounts)
        ? Object.entries(object.accounts).reduce<{ [key: string]: SubscribeRequestFilterAccounts }>(
          (acc, [key, value]) => {
            acc[key] = SubscribeRequestFilterAccounts.fromJSON(value);
            return acc;
          },
          {},
        )
        : {},
      ping: isSet(object.ping) ? SubscribeRequestPing.fromJSON(object.ping) : undefined,
    };
  },

  toJSON(message: SubscribeRequest): unknown {
    const obj: any = {};
    if (message.transactions) {
      const entries = Object.entries(message.transactions);
      if (entries.length > 0) {
        obj.transactions = {};
        entries.forEach(([k, v]) => {
          obj.transactions[k] = SubscribeRequestFilterTransactions.toJSON(v);
        });
      }
    }
    if (message.accounts) {
      const entries = Object.entries(message.accounts);
      if (entries.length > 0) {
        obj.accounts = {};
        entries.forEach(([k, v]) => {
          obj.accounts[k] = SubscribeRequestFilterAccounts.toJSON(v);
        });
      }
    }
    if (message.ping !== undefined) {
      obj.ping = SubscribeRequestPing.toJSON(message.ping);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<SubscribeRequest>, I>>(base?: I): SubscribeRequest {
    return SubscribeRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<SubscribeRequest>, I>>(object: I): SubscribeRequest {
    const message = createBaseSubscribeRequest();
    message.transactions = Object.entries(object.transactions ?? {}).reduce<
      { [key: string]: SubscribeRequestFilterTransactions }
    >((acc, [key, value]) => {
      if (value !== undefined) {
        acc[key] = SubscribeRequestFilterTransactions.fromPartial(value);
      }
      return acc;
    }, {});
    message.accounts = Object.entries(object.accounts ?? {}).reduce<{ [key: string]: SubscribeRequestFilterAccounts }>(
      (acc, [key, value]) => {
        if (value !== undefined) {
          acc[key] = SubscribeRequestFilterAccounts.fromPartial(value);
        }
        return acc;
      },
      {},
    );
    message.ping = (object.ping !== undefined && object.ping !== null)
      ? SubscribeRequestPing.fromPartial(object.ping)
      : undefined;
    return message;
  },
};
// ...
export const SubscribeParsedRequest = {
  encode(message: SubscribeParsedRequest, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.ping !== undefined) {
      SubscribeRequestPing.encode(message.ping, writer.uint32(10).fork()).ldelim();
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): SubscribeParsedRequest {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseSubscribeParsedRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 10) {
            break;
          }

          message.ping = SubscribeRequestPing.decode(reader, reader.uint32());
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): SubscribeParsedRequest {
    return { ping: isSet(object.ping) ? SubscribeRequestPing.fromJSON(object.ping) : undefined };
  },

  toJSON(message: SubscribeParsedRequest): unknown {
    const obj: any = {};
    if (message.ping !== undefined) {
      obj.ping = SubscribeRequestPing.toJSON(message.ping);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<SubscribeParsedRequest>, I>>(base?: I): SubscribeParsedRequest {
    return SubscribeParsedRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<SubscribeParsedRequest>, I>>(object: I): SubscribeParsedRequest {
    const message = createBaseSubscribeParsedRequest();
    message.ping = (object.ping !== undefined && object.ping !== null)
      ? SubscribeRequestPing.fromPartial(object.ping)
      : undefined;
    return message;
  },
};
// ...
export const PingRequest = {
  encode(message: PingRequest, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.count !== 0) {
      writer.uint32(8).int32(message.count);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): PingRequest {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBasePingRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 8) {
            break;
          }

          message.count = reader.int32();
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): PingRequest {
    return { count: isSet(object.count) ? globalThis.Number(object.count) : 0 };
  },

  toJSON(message: PingRequest): unknown {
    const obj: any = {};
    if (message.count !== 0) {
      obj.count = Math.round(message.count);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<PingRequest>, I>>(base?: I): PingRequest {
    return PingRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<PingRequest>, I>>(object: I): PingRequest {
    const message = createBasePingRequest();
    message.count = object.count ?? 0;
    return message;
  },
};
// ...
export type JetstreamService = typeof JetstreamService;
// ...
export interface JetstreamClient extends Client {
  /** Subscribe to data streams with filtering support */
  subscribe(): ClientDuplexStream<SubscribeRequest, SubscribeUpdate>;
  subscribe(options: Partial<CallOptions>): ClientDuplexStream<SubscribeRequest, SubscribeUpdate>;
  subscribe(metadata: Metadata, options?: Partial<CallOptions>): ClientDuplexStream<SubscribeRequest, SubscribeUpdate>;
  /** Subscribe to data streams with filtering support and parsed instructions */
  subscribeParsed(): ClientDuplexStream<SubscribeParsedRequest, SubscribeUpdateParsedTransaction>;
  subscribeParsed(
    options: Partial<CallOptions>,
  ): ClientDuplexStream<SubscribeParsedRequest, SubscribeUpdateParsedTransaction>;
  subscribeParsed(
    metadata: Metadata,
    options?: Partial<CallOptions>,
  ): ClientDuplexStream<SubscribeParsedRequest, SubscribeUpdateParsedTransaction>;
  /** Basic ping/pong for connection testing */
  ping(request: PingRequest, callback: (error: ServiceError | null, response: PongResponse) => void): ClientUnaryCall;
  ping(
    request: PingRequest,
    metadata: Metadata,
    callback: (error: ServiceError | null, response: PongResponse) => void,
  ): ClientUnaryCall;
  ping(
    request: PingRequest,
    metadata: Metadata,
    options: Partial<CallOptions>,
    callback: (error: ServiceError | null, response: PongResponse) => void,
  ): ClientUnaryCall;
  /** Get information about current state */
  getVersion(
    request: GetVersionRequest,
    callback: (error: ServiceError | null, response: GetVersionResponse) => void,
  ): ClientUnaryCall;
  getVersion(
    request: GetVersionRequest,
    metadata: Metadata,
    callback: (error: ServiceError | null, response: GetVersionResponse) => void,
  ): ClientUnaryCall;
  getVersion(
    request: GetVersionRequest,
    metadata: Metadata,
    options: Partial<CallOptions>,
    callback: (error: ServiceError | null, response: GetVersionResponse) => void,
  ): ClientUnaryCall;
}

export const JetstreamClient = makeGenericClientConstructor(JetstreamService, "jetstream.Jetstream") as unknown as {
  new (address: string, credentials: ChannelCredentials, options?: Partial<ClientOptions>): JetstreamClient;
  service: typeof JetstreamService;
  serviceName: string;
};